BE/requirements.txt
requests
pycryptodome
curl_cffi
aiohttp
certifi
browser_cookie3
websockets
js2py
quickjs
flask
flask-cors
django
django-cors-headers
djangorestframework